Refactor AI providers: extract shared logic into Core module
Extract duplicated chain-building, tool construction, message handling,
and response extraction into Lotus.AI.Providers.Core. Add query context
support and source-agnostic Variables prompt module. Fix message type
consistency in Conversation.
